+++ packages/filid/src/hooks/utils/get-parent-segments.ts
@@ -1,0 +1,5 @@
+export function getParentSegments(filePath: string): string[] {
+  const normalized = filePath.replace(/\\/g, '/');
+  const parts = normalized.split('/').filter((p) => p.length > 0);
+  return parts.slice(0, -1);
+}
